+++ haul-be/src/test/java/com/hansalchai/haul/common/utils/KaKaoMapTest.java
@@ -18,8 +18,8 @@ public class KaKaoMapTest {
 	@DisplayName("unittest : 카카오맵 API가 정상적으로 동작한다.")
 	void carPathFindTest() {
 		//given
-		MapUtils.Location src = new MapUtils.Location(BigDecimal.valueOf(127.11015314141542),BigDecimal.valueOf(37.39472714688412));
-		MapUtils.Location dst = new MapUtils.Location(BigDecimal.valueOf(127.10824367964793),BigDecimal.valueOf(37.401937080111644));
+		MapUtils.Location src = new MapUtils.Location(127.11015314141542,37.39472714688412);
+		MapUtils.Location dst = new MapUtils.Location(127.10824367964793,37.401937080111644);
 		double val1 = 1.033;//km
 		int val2 = 254;
 		MapUtils.DistanceDurationInfo expected = new MapUtils.DistanceDurationInfo(val1, val2);
